+++ drizzle/0006_quick_patch.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "wishlist_items" ADD COLUMN "starred" boolean DEFAULT false NOT NULL;